+++ src/Rules/AlphaNum.php
@@ -14,6 +14,6 @@ class AlphaNum extends Rule
             return false;
         }
 
-        return preg_match('/^[\pL\pM\pN]+$/u', $value) > 0;
+        return preg_match('/^[\pL\pM\pN]+$/u', (string)$value) > 0;
     }
 }