+++ tests/Rules/AlphaTest.php
@@ -3,11 +3,13 @@
 namespace Somnambulist\Components\Validation\Tests\Rules;
 
 use PHPUnit\Framework\TestCase;
+use Somnambulist\Components\Validation\Rule;
 use Somnambulist\Components\Validation\Rules\Alpha;
 use stdClass;
 
 class AlphaTest extends TestCase
 {
+    private Rule $rule;
 
     public function setUp(): void
     {