+++ tests/Rules/IntegerTest.php
@@ -3,10 +3,12 @@
 namespace Somnambulist\Components\Validation\Tests\Rules;
 
 use PHPUnit\Framework\TestCase;
+use Somnambulist\Components\Validation\Rule;
 use Somnambulist\Components\Validation\Rules\TypeInteger;
 
 class IntegerTest extends TestCase
 {
+    private Rule $rule;
 
     public function setUp(): void
     {